# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Member Management Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Member {
  id: number;
  registrationNumber: string;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  status: MemberStatus;
  joinDate?: string;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email?: string;
  address?: string;
  notes?: string;
  joinDate?: string;
}

export type MemberUpdateStatus = typeof MemberUpdateStatus[keyof typeof MemberUpdateStatus];


export const MemberUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface MemberUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  status?: MemberUpdateStatus;
  joinDate?: string;
}

export interface Payment {
  id: number;
  memberId: number;
  memberName?: string;
  memberRegistrationNumber?: string;
  amount: number;
  month: number;
  year: number;
  /** @nullable */
  notes?: string | null;
  paidAt: string;
}

export interface PaymentInput {
  memberId: number;
  amount: number;
  month: number;
  year: number;
  notes?: string;
}

export interface MemberPaymentStatus {
  memberId: number;
  registrationNumber: string;
  name: string;
  phone: string;
  paid: boolean;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  paymentId?: number | null;
  month: number;
  year: number;
}

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  paidThisMonth: number;
  unpaidThisMonth: number;
  totalCollectedThisMonth: number;
  totalCollectedThisYear: number;
  month: number;
  year: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  totalMembers: number;
  paid: number;
  unpaid: number;
  totalCollected: number;
  expectedTotal: number;
  collectionRate: number;
  payments: MemberPaymentStatus[];
}

export interface MonthSummary {
  month: number;
  year: number;
  paid: number;
  unpaid: number;
  collected: number;
}

export interface YearlyReport {
  year: number;
  totalCollected: number;
  totalExpected: number;
  monthlyBreakdown: MonthSummary[];
}

export interface Settings {
  id: number;
  organizationName: string;
  monthlyDueAmount: number;
  whatsappAlertTemplate: string;
  currency: string;
}

export interface SettingsUpdate {
  organizationName?: string;
  monthlyDueAmount?: number;
  whatsappAlertTemplate?: string;
  currency?: string;
}

export type ListMembersParams = {
search?: string;
status?: ListMembersStatus;
};

export type ListMembersStatus = typeof ListMembersStatus[keyof typeof ListMembersStatus];


export const ListMembersStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListPaymentsParams = {
month?: number;
year?: number;
memberId?: number;
};

export type GetPaymentStatusParams = {
month: number;
year: number;
};

export type GetDashboardStatsParams = {
month?: number;
year?: number;
};

export type GetMonthlyReportParams = {
month: number;
year: number;
};

export type GetYearlyReportParams = {
year: number;
};

export type GetUnpaidMembersParams = {
month: number;
year: number;
};