# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Member Management Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  paidThisMonth: number;
  unpaidThisMonth: number;
  totalCollectedThisMonth: number;
  totalCollectedThisYear: number;
  month: number;
  year: number;
}