# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Member Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateStatus } from './memberUpdateStatus';

export interface MemberUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  status?: MemberUpdateStatus;
  joinDate?: string;
}